ditto-markov/pyproject.toml
[tool.poetry]
name = "ditto-markov"
version = "0.1.0"
description = ""
authors = ["Ridwan Hoq <ridwanhoq@gmail.com>"]

[tool.poetry.scripts]
cli = "ditto_markov.cli:main"

[tool.poetry.dependencies]
python = "^3.7"
azure-cosmos = "^3.1.2"
markovify = "^0.8.0"
azure-storage-blob = "^12.3.1"
cachetools = "^4.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
